// src/features/editableProfileCard/model/const/consts.ts
export enum ValidateProfileError {
  INCORRECT_USERNAME = 'INCORRECT_USERNAME',
  INCORRECT_LASTNAME = 'INCORRECT_LASTNAME',
  INCORRECT_FIRSTNAME = 'INCORRECT_FIRSTNAME',
  INCORRECT_CITY = 'INCORRECT_CITY',
  INCORRECT_AVATAR = 'INCORRECT_AVATAR',
  INCORRECT_COUNTRY = 'INCORRECT_COUNTRY',
  INCORRECT_CURRENCY = 'INCORRECT_CURRENCY',
  INCORRECT_AGE = 'INCORRECT_AGE',
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
  NO_DATA = 'NO_DATA'
}